src/requirements.txt
gunicorn==19.7.1
whitenoise==5.0
livereload
Django>=2.2.8
djangorestframework
django-allauth
numpy==1.19.5
tensorflow-cpu==2.2.0
